# server/api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: Vector API
  description: Full API specification for the Vector fleet management platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /dashboard/orders:
    get:
      summary: List all orders
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of orders with pagination
    post:
      summary: Create new order/stop
      tags: [Dashboard]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Order created
  /dashboard/orders/bulk:
    post:
      summary: Bulk import orders
      tags: [Dashboard]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                orders:
                  type: array
                  items:
                    $ref: '#/components/schemas/CreateOrderDto'
      responses:
        '201':
          description: Bulk import result
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                    example: 42
                  skipped:
                    type: integer
                    example: 0
                  errors:
                    type: array
                    items:
                      type: string
  /dashboard/orders/{stop_id}:
    patch:
      summary: Update order
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: stop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              # Add update fields here
      responses:
        '200':
          description: Order updated
    delete:
      summary: Delete order
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: stop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Order deleted
...
  /dashboard/reports/summary:
    get:
      summary: Get report summary metrics
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Report summary
  /dashboard/reports/charts:
    get:
      summary: Get data for reports charts
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Chart data
  /dashboard/reports/performance:
    get:
      summary: Get driver performance
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Paginated driver performance stats
  /dashboard/reports/export:
    get:
      summary: Export report as CSV
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: CSV file attachment
          content:
            text/csv:
              schema:
                type: string
  /dashboard/billing:
    get:
      summary: Get billing info
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Billing plan and status
  /dashboard/billing/invoices:
    get:
      summary: Get invoice history
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of invoices
...
  /dashboard/billing/cancel:
    post:
      summary: Cancel subscription
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Subscription cancelled
  /dashboard/settings:
    get:
      summary: Get company settings
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings data
    patch:
      summary: Update settings
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings updated
  /dashboard/settings/api-keys:
    post:
      summary: Create new API key
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '201':
          description: API key generated
  /dashboard/settings/api-keys/{key_id}:
    delete:
      summary: Revoke API key
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: key_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Key revoked
...
  /dashboard/notifications/{notification_id}/read:
    patch:
      summary: Mark notification as read
      tags: [Dashboard]
      security:
        - bearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marked as read
  /dashboard/notifications/read-all:
    post:
      summary: Mark all as read
      tags: [Dashboard]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All marked read
...
  /driver/navigation/stops/{stop_id}:
    get:
      summary: Get stop detail
      tags: [Driver]
      security:
        - bearerAuth: []
      parameters:
        - name: stop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Stop detail
  /driver/navigation/stops/{stop_id}/arrive:
    patch:
      summary: Mark manual arrival at stop
      tags: [Driver]
      security:
        - bearerAuth: []
      parameters:
        - name: stop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marked arrived
  /driver/navigation/stops/{stop_id}/skip:
    patch:
      summary: Skip a stop
      tags: [Driver]
      security:
        - bearerAuth: []
      parameters:
        - name: stop_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marked skipped/failed
...
  /driver/profile:
    get:
      summary: Get driver profile
      tags: [Driver]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile retrieved
    patch:
      summary: Update driver profile
      tags: [Driver]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated
  /driver/profile/avatar:
    post:
      summary: Upload new avatar
      tags: [Driver]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Avatar updated
...
  /driver/notifications:
    get:
      summary: Driver notifications
      tags: [Driver]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification list
  /driver/notifications/{notification_id}/read:
    patch:
      summary: Mark read
      tags: [Driver]
      security:
        - bearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Marked read
  /driver/notifications/read-all:
    post:
      summary: Mark all read
      tags: [Driver]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All marked read
...
  /track:
    get:
      summary: Track delivery status
      tags: [Tracking]
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Public delivery info
  /track/rate:
    post:
      summary: Submit delivery rating
      tags: [Tracking]
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                feedback:
                  type: string
      responses:
        '200':
          description: Rating received
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateOrderDto:
      type: object
      required:
        - customer_name
        - address
        - is_priority
      properties:
        customer_name:
          type: string
          example: John Doe
        address:
          type: string
          example: 123 Main St
        lat:
          type: number
          example: 40.7128
        lng:
          type: number
          example: -74.0060
        is_priority:
          type: boolean
          example: false
...